src/proof.rs
use bellman::{
    groth16::{
        create_random_proof, generate_random_parameters, prepare_verifying_key, verify_proof,
        Parameters, Proof, VerifyingKey,
    },
    Circuit,
};
use ff::PrimeField;
use rand::thread_rng;
use sha3::{Digest, Sha3_256};
use std::sync::Arc;
use parking_lot::RwLock;
use rayon::prelude::*;
use blake2::{Blake2b512, Blake2s256};

#[derive(Clone)]
pub struct ProofSystem<F: PrimeField> {
    params: Arc<Parameters<F>>,
    verifying_key: Arc<VerifyingKey<F>>,
    proof_cache: Arc<RwLock<lru::LruCache<[u8; 32], Proof<F>>>>,
}

impl<F: PrimeField> ProofSystem<F> {
    pub fn setup<C: Circuit<F>>(circuit: C) -> Result<Self, Box<dyn std::error::Error>> {
        let rng = &mut thread_rng();
        let params = generate_random_parameters::<F, _, _>(circuit, rng)?;
        let verifying_key = params.vk.clone();
        
        Ok(Self {
            params: Arc::new(params),
            verifying_key: Arc::new(verifying_key),
            proof_cache: Arc::new(RwLock::new(lru::LruCache::new(1000))),
        })
    }

    pub fn prove<C: Circuit<F>>(&self, circuit: C) -> Result<ProofData<F>, Box<dyn std::error::Error>> {
        let rng = &mut thread_rng();
        
        // Generate proof
        let proof = create_random_proof(circuit, &self.params, rng)?;
        
        // Collect public inputs
        let public_inputs = self.collect_public_inputs(&proof)?;
        
        // Generate proof hash
        let proof_hash = self.hash_proof(&proof, &public_inputs)?;
        
        // Cache the proof
        self.proof_cache.write().put(proof_hash, proof.clone());
        
        Ok(ProofData::new(proof, public_inputs, proof_hash))
    }

    pub fn verify(&self, proof_data: &ProofData<F>) -> Result<bool, Box<dyn std::error::Error>> {
        // Check cache first
        if let Some(cached_proof) = self.proof_cache.read().get(&proof_data.hash) {
            if cached_proof == &proof_data.proof {
                return Ok(true);
            }
        }
        
        // Verify proof
        let pvk = prepare_verifying_key(&self.verifying_key);
        let is_valid = verify_proof(&pvk, &proof_data.proof, &proof_data.public_inputs)?;
        
        // Verify hash
        let computed_hash = self.hash_proof(&proof_data.proof, &proof_data.public_inputs)?;
        let hash_valid = computed_hash == proof_data.hash;
        
        Ok(is_valid && hash_valid)
    }

    pub fn batch_verify(&self, proofs: &[ProofData<F>]) -> Result<bool, Box<dyn std::error::Error>> {
        let pvk = prepare_verifying_key(&self.verifying_key);
        
        // Parallel verification
        let results: Vec<bool> = proofs.par_iter().map(|proof_data| {
            // Check cache
            if let Some(cached_proof) = self.proof_cache.read().get(&proof_data.hash) {
                return cached_proof == &proof_data.proof;
            }
            
            // Verify proof
            match verify_proof(&pvk, &proof_data.proof, &proof_data.public_inputs) {
                Ok(is_valid) => is_valid,
                Err(_) => false,
            }
        }).collect();
        
        Ok(results.iter().all(|&x| x))
    }

    fn collect_public_inputs(&self, proof: &Proof<F>) -> Result<Vec<F>, Box<dyn std::error::Error>> {
        // Implementation depends on circuit structure
        // This is a placeholder that should be customized based on your circuit
        Ok(Vec::new())
    }

    fn hash_proof(&self, proof: &Proof<F>, public_inputs: &[F]) -> Result<[u8; 32], Box<dyn std::error::Error>> {
        let mut hasher = Blake2s256::new();
        
        // Hash proof components
        hasher.update(&proof.a.to_uncompressed());
        hasher.update(&proof.b.to_uncompressed());
        hasher.update(&proof.c.to_uncompressed());
        
        // Hash public inputs
        for input in public_inputs {
            hasher.update(&input.to_repr());
        }
        
        let mut hash = [0u8; 32];
        hash.copy_from_slice(&hasher.finalize());
        Ok(hash)
    }
}

#[derive(Clone)]
pub struct ProofData<F: PrimeField> {
    pub proof: Proof<F>,
    pub public_inputs: Vec<F>,
    pub hash: [u8; 32],
}

impl<F: PrimeField> ProofData<F> {
    pub fn new(proof: Proof<F>, public_inputs: Vec<F>, hash: [u8; 32]) -> Self {
        Self {
            proof,
            public_inputs,
            hash,
        }
    }
}

pub struct BatchVerificationError {
    pub index: usize,
    pub error: Box<dyn std::error::Error>,
}

#[derive(Default)]
pub struct ProofSystemBuilder<F: PrimeField> {
    cache_size: Option<usize>,
    parallel_verification: bool,
    verification_batch_size: Option<usize>,
}

impl<F: PrimeField> ProofSystemBuilder<F> {
    pub fn new() -> Self {
        Self {
            cache_size: Some(1000),
            parallel_verification: true,
            verification_batch_size: Some(100),
        }
    }

    pub fn with_cache_size(mut self, size: usize) -> Self {
        self.cache_size = Some(size);
        self
    }

    pub fn with_parallel_verification(mut self, enabled: bool) -> Self {
        self.parallel_verification = enabled;
        self
    }

    pub fn with_verification_batch_size(mut self, size: usize) -> Self {
        self.verification_batch_size = Some(size);
        self
    }

    pub fn build<C: Circuit<F>>(self, circuit: C) -> Result<ProofSystem<F>, Box<dyn std::error::Error>> {
        let rng = &mut thread_rng();
        let params = generate_random_parameters::<F, _, _>(circuit, rng)?;
        let verifying_key = params.vk.clone();
        
        Ok(ProofSystem {
            params: Arc::new(params),
            verifying_key: Arc::new(verifying_key),
            proof_cache: Arc::new(RwLock::new(lru::LruCache::new(
                self.cache_size.unwrap_or(1000)
            ))),
        })
    }
}

#[cfg(test)]
mod tests {
    use super::*;
    use ff::Field;
    use rand::Rng;

    #[test]
    fn test_proof_system() {
        // Add test implementation
    }

    #[test]
    fn test_batch_verification() {
        // Add test implementation
    }

    #[test]
    fn test_proof_caching() {
        // Add test implementation
    }
}
